security: harden MC server with AppArmor and cgroups
AppArmor implements MAC for Linux processes, allowing to restrict their
permissions and capabilities to a subset of those that would have via
the usual DAC mechanisms (Unix users, groups and setuid binaries). This
adds another security layer that helps preventing the effects of rogue
plugins and zero-day exploits. Best of all, AppArmor is already bundled
with Ubuntu, so this MAC comes for free, unlike containers, which
introduce unnecessary isolation and overhead.

While at it, leverage systemd service cgroup resource limitations to
limit the listen sockets and memory consumption of the server, which is
useful for security and resource planning purposes.