fix(quotas): make them work for good by using proper kernels and FS opts
For reasons unbeknownst to me, the Oracle Cloud kernel images stopped
shipping the quota kernel modules that are part of more standard Ubuntu
image packages, rendering storage quotas non-functional for a while
(this is something I tried to work around in the past by pinning the
kernel image to a known-good version, in case it was a temporary fluke
in how the package was generated, but it looks like it was not).

Moreover, using external quota files as we were doing is deprecated
(https://askubuntu.com/questions/1492000/deprecation-warning-for-quota-how-to-fix),
but the alternative of enabling the ext4 filesystem quota option was
very inconvenient, as such filesystem was the root filesystem, which
cannot be unmounted from the instance itself while it's booted to change
it.

Thanks to the power of testing stuff with sacrificial instances and VMs,
which took a while in big part due to Oracle's available capacity for
these being low, I now feel confident enough to use a initramfs boot
script to set the desired root filesystem options during early boot
stages, before it's mounted there for good. I also feel confident enough
to use generic kernel images with the required modules. Therefore,
storage quota issues should now be fixed for good, finally 🎉