fix(rustic_backup): make mirroring to MEGA reliably work
MegaCMD's WebDAV server has also proven to be troublesome when using
rustic's OpenDAL integration, as it frequently returns spurious errors
that may cause apparent or real repository corruption. I've tested a few
other WebDAV clients, but all of them had pretty bad performance anyway,
not fully utilizing the network bandwidth available by a long stretch,
so the most likely conclusion is that such WebDAV server is lackluster.

Therefore, as we can't use rclone's MEGA backend, and WebDAV has been
also proven to be unusable, let's fall back to MegaCMD's folder
synchronization feature. Luckily, this approach has been proven to be
the most performant and stable by a wide margin, while straightforward
enough to implement through rustic's hooks. The only downside is that
such synchronization is bidirectional, so we should take care of not
modifying the target MEGA folder in a way that corrupts the backup
state.